chore(tech-radar): promote nfs plugin from ./alpha to primary entry point
