+++ workspaces/tech-radar/app-config.yaml
@@ -50,7 +50,7 @@ catalog:
   rules:
     - allow: [Component, System, API, Resource, Location]
   locations:
-    # Local example data, file locations are relative to the backend process, typically `packages/backend`
+    # Local example data, file locations are relative to the backend process
     - type: file
       target: ../../examples/entities.yaml
