--- workspaces/tech-radar/packages/app-next/knip-report.md
@@ -1,31 +1,0 @@
-# Knip report
-
-## Unused dependencies (16)
-
-| Name                             | Location     | Severity |
-| :------------------------------- | :----------- | :------- |
-| @backstage/plugin-catalog-common | package.json | error    |
-| @backstage/plugin-catalog-graph  | package.json | error    |
-| @backstage/plugin-catalog-react  | package.json | error    |
-| @backstage/frontend-app-api      | package.json | error    |
-| @backstage/core-compat-api       | package.json | error    |
-| @backstage/core-components       | package.json | error    |
-| @backstage/core-plugin-api       | package.json | error    |
-| @backstage/plugin-api-docs       | package.json | error    |
-| @backstage/catalog-model         | package.json | error    |
-| @backstage/app-defaults          | package.json | error    |
-| @backstage/plugin-org            | package.json | error    |
-| @material-ui/icons               | package.json | error    |
-| @material-ui/core                | package.json | error    |
-| styled-components                | package.json | error    |
-| react-use                        | package.json | error    |
-| history                          | package.json | error    |
-
-## Unused devDependencies (4)
-
-| Name                        | Location     | Severity |
-| :-------------------------- | :----------- | :------- |
-| @testing-library/user-event | package.json | error    |
-| @testing-library/dom        | package.json | error    |
-| @playwright/test            | package.json | error    |
-| cross-env                   | package.json | error    |
