--- workspaces/tech-radar/packages/app-next/package.json
@@ -1,78 +1,0 @@
-{
-  "name": "app-next",
-  "version": "0.0.3",
-  "private": true,
-  "bundled": true,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/backstage/community-plugins",
-    "directory": "workspaces/tech-radar/packages/app-next"
-  },
-  "backstage": {
-    "role": "frontend"
-  },
-  "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
-    "clean": "backstage-cli package clean",
-    "test": "backstage-cli package test",
-    "lint": "backstage-cli package lint"
-  },
-  "dependencies": {
-    "@backstage-community/plugin-tech-radar": "workspace:^",
-    "@backstage-community/plugin-tech-radar-common": "workspace:^",
-    "@backstage/app-defaults": "backstage:^",
-    "@backstage/catalog-model": "backstage:^",
-    "@backstage/cli": "backstage:^",
-    "@backstage/core-compat-api": "backstage:^",
-    "@backstage/core-components": "backstage:^",
-    "@backstage/core-plugin-api": "backstage:^",
-    "@backstage/frontend-app-api": "backstage:^",
-    "@backstage/frontend-defaults": "backstage:^",
-    "@backstage/frontend-plugin-api": "backstage:^",
-    "@backstage/plugin-api-docs": "backstage:^",
-    "@backstage/plugin-catalog": "backstage:^",
-    "@backstage/plugin-catalog-common": "backstage:^",
-    "@backstage/plugin-catalog-graph": "backstage:^",
-    "@backstage/plugin-catalog-import": "backstage:^",
-    "@backstage/plugin-catalog-react": "backstage:^",
-    "@backstage/plugin-org": "backstage:^",
-    "@backstage/plugin-signals": "backstage:^",
-    "@backstage/plugin-user-settings": "backstage:^",
-    "@backstage/theme": "backstage:^",
-    "@material-ui/core": "^4.12.2",
-    "@material-ui/icons": "^4.9.1",
-    "history": "^5.0.0",
-    "react": "^18.0.2",
-    "react-dom": "^18.0.2",
-    "react-router": "^7.18.2",
-    "react-router-dom": "^6.3.0",
-    "react-use": "^17.2.4",
-    "styled-components": "^6.1.8"
-  },
-  "devDependencies": {
-    "@backstage/test-utils": "backstage:^",
-    "@playwright/test": "^1.32.3",
-    "@testing-library/dom": "^9.0.0",
-    "@testing-library/jest-dom": "^6.0.0",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.0.0",
-    "@types/react-dom": "*",
-    "cross-env": "^10.0.0"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "files": [
-    "dist"
-  ]
-}
