--- workspaces/tech-radar/packages/app-next/src/App.test.tsx
@@ -1,44 +1,0 @@
-/*
- * Copyright 2024 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { renderWithEffects } from '@backstage/test-utils';
-
-// Rarely, and only in windows CI, do these tests take slightly more than the
-// default five seconds
-jest.setTimeout(15_000);
-
-describe('App', () => {
-  it('should render', async () => {
-    process.env = {
-      NODE_ENV: 'test',
-      APP_CONFIG: [
-        {
-          data: {
-            app: {
-              title: 'Test',
-              support: { url: 'http://localhost:7007/support' },
-            },
-            backend: { baseUrl: 'http://localhost:7007' },
-          },
-          context: 'test',
-        },
-      ] as any,
-    };
-
-    const { default: app } = await import('./App');
-    const rendered = await renderWithEffects(app);
-    expect(rendered.baseElement).toBeInTheDocument();
-  });
-});
