--- workspaces/tech-radar/packages/app/.eslintignore
@@ -1,1 +1,0 @@
-public
