--- workspaces/tech-radar/packages/app/CHANGELOG.md
@@ -1,17 +1,0 @@
-# app
-
-## 0.0.3
-
-### Patch Changes
-
-- Updated dependencies [e3ab8e9]
-- Updated dependencies [e5a9abf]
-  - @backstage-community/plugin-tech-radar-common@1.0.0
-  - @backstage-community/plugin-tech-radar@1.0.0
-
-## 0.0.2
-
-### Patch Changes
-
-- Updated dependencies [6e53587]
-  - @backstage-community/plugin-tech-radar@0.7.11
