--- workspaces/tech-radar/packages/app/knip-report.md
@@ -1,7 +1,0 @@
-# Knip report
-
-## Unused devDependencies (1)
-
-| Name             | Location     | Severity |
-| :--------------- | :----------- | :------- |
-| @playwright/test | package.json | error    |
