--- workspaces/tech-radar/packages/app/package.json
@@ -1,74 +1,0 @@
-{
-  "name": "app",
-  "version": "0.0.3",
-  "private": true,
-  "bundled": true,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/backstage/community-plugins",
-    "directory": "workspaces/tech-radar/packages/app"
-  },
-  "backstage": {
-    "role": "frontend"
-  },
-  "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
-    "clean": "backstage-cli package clean",
-    "test": "backstage-cli package test",
-    "lint": "backstage-cli package lint"
-  },
-  "dependencies": {
-    "@backstage-community/plugin-tech-radar": "workspace:^",
-    "@backstage-community/plugin-tech-radar-common": "workspace:^",
-    "@backstage/app-defaults": "backstage:^",
-    "@backstage/catalog-model": "backstage:^",
-    "@backstage/cli": "backstage:^",
-    "@backstage/core-app-api": "backstage:^",
-    "@backstage/core-components": "backstage:^",
-    "@backstage/core-plugin-api": "backstage:^",
-    "@backstage/integration-react": "backstage:^",
-    "@backstage/plugin-api-docs": "backstage:^",
-    "@backstage/plugin-catalog": "backstage:^",
-    "@backstage/plugin-catalog-common": "backstage:^",
-    "@backstage/plugin-catalog-graph": "backstage:^",
-    "@backstage/plugin-catalog-import": "backstage:^",
-    "@backstage/plugin-catalog-react": "backstage:^",
-    "@backstage/plugin-org": "backstage:^",
-    "@backstage/plugin-permission-react": "backstage:^",
-    "@backstage/plugin-scaffolder": "backstage:^",
-    "@backstage/plugin-search": "backstage:^",
-    "@backstage/plugin-search-react": "backstage:^",
-    "@backstage/plugin-techdocs": "backstage:^",
-    "@backstage/plugin-techdocs-module-addons-contrib": "backstage:^",
-    "@backstage/plugin-techdocs-react": "backstage:^",
-    "@backstage/plugin-user-settings": "backstage:^",
-    "@backstage/theme": "backstage:^",
-    "@material-ui/core": "^4.12.2",
-    "@material-ui/icons": "^4.9.1",
-    "react": "^18.0.2",
-    "react-dom": "^18.0.2",
-    "react-router-dom": "^6.3.0"
-  },
-  "devDependencies": {
-    "@playwright/test": "^1.32.3",
-    "@testing-library/jest-dom": "^6.0.0",
-    "@testing-library/react": "^14.0.0",
-    "@types/react-dom": "*"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "files": [
-    "dist"
-  ]
-}
