--- workspaces/tech-radar/packages/app/src/apis.ts
@@ -1,58 +1,0 @@
-/*
- * Copyright 2024 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import {
-  ScmIntegrationsApi,
-  scmIntegrationsApiRef,
-  ScmAuth,
-} from '@backstage/integration-react';
-import {
-  AnyApiFactory,
-  configApiRef,
-  createApiFactory,
-} from '@backstage/core-plugin-api';
-import {
-  TechRadarApi,
-  // @ts-ignore
-  techRadarApiRef,
-} from '@backstage-community/plugin-tech-radar';
-import { TechRadarLoaderResponse } from '@backstage-community/plugin-tech-radar-common';
-
-// overriding the api is one way to change the radar content
-// @ts-ignore
-class SampleTechRadarApi implements TechRadarApi {
-  async load(): Promise<TechRadarLoaderResponse> {
-    return {
-      entries: [],
-      quadrants: [
-        { id: 'infrastructure', name: 'Infrastructure' },
-        { id: 'frameworks', name: 'Frameworks' },
-        { id: 'languages', name: 'Languages' },
-        { id: 'process', name: 'Process' },
-      ],
-      rings: [],
-    };
-  }
-}
-
-export const apis: AnyApiFactory[] = [
-  createApiFactory({
-    api: scmIntegrationsApiRef,
-    deps: { configApi: configApiRef },
-    factory: ({ configApi }) => ScmIntegrationsApi.fromConfig(configApi),
-  }),
-  ScmAuth.createDefaultApiFactory(),
-  // createApiFactory(techRadarApiRef, new SampleTechRadarApi()), // comment out to use a custom api
-];
