--- workspaces/tech-radar/packages/backend/CHANGELOG.md
@@ -1,10 +1,0 @@
-# backend
-
-## 0.0.2
-
-### Patch Changes
-
-- Updated dependencies [e3ab8e9]
-- Updated dependencies [e3ab8e9]
-  - @backstage-community/plugin-tech-radar-common@1.0.0
-  - @backstage-community/plugin-tech-radar-backend@1.0.0
