--- workspaces/tech-radar/packages/backend/knip-report.md
@@ -1,11 +1,0 @@
-# Knip report
-
-## Unused dependencies (5)
-
-| Name                                                  | Location     | Severity |
-| :---------------------------------------------------- | :----------- | :------- |
-| @backstage/plugin-auth-backend-module-github-provider | package.json | error    |
-| @backstage-community/plugin-tech-radar-common         | package.json | error    |
-| @backstage/backend-plugin-api                         | package.json | error    |
-| @backstage/plugin-auth-node                           | package.json | error    |
-| @backstage/catalog-model                              | package.json | error    |
