--- workspaces/tech-radar/packages/backend/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "backend",
-  "version": "0.0.2",
-  "main": "dist/index.cjs.js",
-  "types": "src/index.ts",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/backstage/community-plugins",
-    "directory": "workspaces/tech-radar/packages/backend"
-  },
-  "backstage": {
-    "role": "backend"
-  },
-  "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
-    "lint": "backstage-cli package lint",
-    "test": "backstage-cli package test",
-    "clean": "backstage-cli package clean",
-    "build-image": "docker build ../.. -f Dockerfile --tag backstage"
-  },
-  "dependencies": {
-    "@backstage-community/plugin-tech-radar-backend": "workspace:^",
-    "@backstage-community/plugin-tech-radar-common": "workspace:^",
-    "@backstage/backend-defaults": "backstage:^",
-    "@backstage/backend-plugin-api": "backstage:^",
-    "@backstage/catalog-model": "backstage:^",
-    "@backstage/plugin-app-backend": "backstage:^",
-    "@backstage/plugin-auth-backend": "backstage:^",
-    "@backstage/plugin-auth-backend-module-github-provider": "backstage:^",
-    "@backstage/plugin-auth-backend-module-guest-provider": "backstage:^",
-    "@backstage/plugin-auth-node": "backstage:^",
-    "@backstage/plugin-catalog-backend": "backstage:^",
-    "@backstage/plugin-permission-backend": "backstage:^",
-    "@backstage/plugin-permission-backend-module-allow-all-policy": "backstage:^",
-    "@backstage/plugin-proxy-backend": "backstage:^",
-    "@backstage/plugin-scaffolder-backend": "backstage:^",
-    "@backstage/plugin-techdocs-backend": "backstage:^"
-  },
-  "devDependencies": {
-    "@backstage/cli": "backstage:^"
-  },
-  "files": [
-    "dist"
-  ]
-}
