+++ workspaces/tech-radar/plugins/tech-radar-backend/package.json
@@ -26,6 +26,7 @@
   },
   "scripts": {
     "start": "backstage-cli package start",
+    "start:legacy": "backstage-cli package start",
     "build": "backstage-cli package build",
     "lint": "backstage-cli package lint",
     "test": "backstage-cli package test",
