+++ workspaces/tech-radar/plugins/tech-radar/dev/index.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2026 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,68 @@
  * limitations under the License.
  */
 
-import { createDevApp } from '@backstage/dev-utils';
-import { techRadarPlugin, TechRadarPage } from '../src';
+import ReactDOM from 'react-dom/client';
 
-createDevApp()
-  .registerPlugin(techRadarPlugin)
-  .addPage({
-    title: 'Tech Radar',
-    element: <TechRadarPage />,
-  })
-  .render();
+// eslint-disable-next-line @backstage/no-ui-css-imports-in-non-frontend
+import '@backstage/ui/css/styles.css';
+
+import { createApp } from '@backstage/frontend-defaults';
+import {
+  ApiBlueprint,
+  createFrontendModule,
+} from '@backstage/frontend-plugin-api';
+import catalogPlugin from '@backstage/plugin-catalog/alpha';
+import { catalogApiRef } from '@backstage/plugin-catalog-react';
+import { catalogApiMock } from '@backstage/plugin-catalog-react/testUtils';
+import userSettingsPlugin from '@backstage/plugin-user-settings/alpha';
+
+import techRadarPlugin from '../src';
+
+const guestUser = {
+  apiVersion: 'backstage.io/v1alpha1',
+  kind: 'User',
+  metadata: {
+    name: 'guest',
+  },
+  spec: {
+    profile: {
+      displayName: 'Guest User',
+      email: 'guest@example.com',
+    },
+    memberOf: ['guests'],
+  },
+};
+
+const catalogDevModule = createFrontendModule({
+  pluginId: 'catalog',
+  extensions: [
+    ApiBlueprint.make({
+      params: defineParams =>
+        defineParams({
+          api: catalogApiRef,
+          deps: {},
+          factory: () =>
+            catalogApiMock({
+              entities: [guestUser],
+            }),
+        }),
+    }),
+  ],
+});
+
+const defaultPage = '/tech-radar';
+
+const app = createApp({
+  features: [
+    catalogPlugin,
+    catalogDevModule,
+    userSettingsPlugin,
+    techRadarPlugin,
+  ],
+});
+
+if (typeof window !== 'undefined' && window.location.pathname === '/') {
+  window.location.pathname = defaultPage;
+}
+
+ReactDOM.createRoot(document.getElementById('root')!).render(app.createRoot());
