--- workspaces/tech-radar/plugins/tech-radar/report-alpha.api.md
@@ -1,233 +1,0 @@
-## API Report File for "@backstage-community/plugin-tech-radar"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-import { AnyApiFactory } from '@backstage/frontend-plugin-api';
-import { AnyRouteRefParams } from '@backstage/frontend-plugin-api';
-import { ApiFactory } from '@backstage/frontend-plugin-api';
-import { ConfigurableExtensionDataRef } from '@backstage/frontend-plugin-api';
-import { ExtensionBlueprintParams } from '@backstage/frontend-plugin-api';
-import { ExtensionDataRef } from '@backstage/frontend-plugin-api';
-import { ExtensionInput } from '@backstage/frontend-plugin-api';
-import { IconElement } from '@backstage/frontend-plugin-api';
-import { JSX as JSX_2 } from 'react';
-import { OverridableExtensionDefinition } from '@backstage/frontend-plugin-api';
-import { OverridableFrontendPlugin } from '@backstage/frontend-plugin-api';
-import { RouteRef } from '@backstage/frontend-plugin-api';
-
-// @alpha (undocumented)
-const _default: OverridableFrontendPlugin<
-  {
-    root: RouteRef<undefined>;
-  },
-  {},
-  {
-    'api:tech-radar': OverridableExtensionDefinition<{
-      kind: 'api';
-      name: undefined;
-      config: {};
-      configInput: {};
-      output: ExtensionDataRef<AnyApiFactory, 'core.api.factory', {}>;
-      inputs: {};
-      params: <
-        TApi,
-        TImpl extends TApi,
-        TDeps extends { [name in string]: unknown },
-      >(
-        params: ApiFactory<TApi, TImpl, TDeps>,
-      ) => ExtensionBlueprintParams<AnyApiFactory>;
-    }>;
-    'page:tech-radar': OverridableExtensionDefinition<{
-      config: {
-        subtitle: string;
-        pageTitle: string;
-        width: number | undefined;
-        height: number | undefined;
-        path: string | undefined;
-        title: string | undefined;
-      };
-      configInput: {
-        subtitle?: string | undefined;
-        pageTitle?: string | undefined;
-        width?: number | undefined;
-        height?: number | undefined;
-        path?: string | undefined | undefined;
-        title?: string | undefined | undefined;
-      };
-      output:
-        | ExtensionDataRef<JSX_2.Element, 'core.reactElement', {}>
-        | ExtensionDataRef<string, 'core.routing.path', {}>
-        | ExtensionDataRef<
-            RouteRef<AnyRouteRefParams>,
-            'core.routing.ref',
-            {
-              optional: true;
-            }
-          >
-        | ExtensionDataRef<
-            string,
-            'core.title',
-            {
-              optional: true;
-            }
-          >
-        | ExtensionDataRef<
-            IconElement,
-            'core.icon',
-            {
-              optional: true;
-            }
-          >;
-      inputs: {
-        pages: ExtensionInput<
-          | ConfigurableExtensionDataRef<JSX_2.Element, 'core.reactElement', {}>
-          | ConfigurableExtensionDataRef<string, 'core.routing.path', {}>
-          | ConfigurableExtensionDataRef<
-              RouteRef<AnyRouteRefParams>,
-              'core.routing.ref',
-              {
-                optional: true;
-              }
-            >
-          | ConfigurableExtensionDataRef<
-              string,
-              'core.title',
-              {
-                optional: true;
-              }
-            >
-          | ConfigurableExtensionDataRef<
-              IconElement,
-              'core.icon',
-              {
-                optional: true;
-              }
-            >,
-          {
-            singleton: false;
-            optional: false;
-            internal: false;
-          }
-        >;
-      };
-      kind: 'page';
-      name: undefined;
-      params: {
-        path: string;
-        title?: string;
-        icon?: IconElement;
-        loader?: () => Promise<JSX_2.Element>;
-        routeRef?: RouteRef;
-        noHeader?: boolean;
-      };
-    }>;
-  }
->;
-export default _default;
-
-// @alpha (undocumented)
-export const techRadarApi: OverridableExtensionDefinition<{
-  kind: 'api';
-  name: undefined;
-  config: {};
-  configInput: {};
-  output: ExtensionDataRef<AnyApiFactory, 'core.api.factory', {}>;
-  inputs: {};
-  params: <
-    TApi,
-    TImpl extends TApi,
-    TDeps extends { [name in string]: unknown },
-  >(
-    params: ApiFactory<TApi, TImpl, TDeps>,
-  ) => ExtensionBlueprintParams<AnyApiFactory>;
-}>;
-
-// @alpha (undocumented)
-export const techRadarPage: OverridableExtensionDefinition<{
-  config: {
-    subtitle: string;
-    pageTitle: string;
-    width: number | undefined;
-    height: number | undefined;
-    path: string | undefined;
-    title: string | undefined;
-  };
-  configInput: {
-    subtitle?: string | undefined;
-    pageTitle?: string | undefined;
-    width?: number | undefined;
-    height?: number | undefined;
-    path?: string | undefined | undefined;
-    title?: string | undefined | undefined;
-  };
-  output:
-    | ExtensionDataRef<JSX_2.Element, 'core.reactElement', {}>
-    | ExtensionDataRef<string, 'core.routing.path', {}>
-    | ExtensionDataRef<
-        RouteRef<AnyRouteRefParams>,
-        'core.routing.ref',
-        {
-          optional: true;
-        }
-      >
-    | ExtensionDataRef<
-        string,
-        'core.title',
-        {
-          optional: true;
-        }
-      >
-    | ExtensionDataRef<
-        IconElement,
-        'core.icon',
-        {
-          optional: true;
-        }
-      >;
-  inputs: {
-    pages: ExtensionInput<
-      | ConfigurableExtensionDataRef<JSX_2.Element, 'core.reactElement', {}>
-      | ConfigurableExtensionDataRef<string, 'core.routing.path', {}>
-      | ConfigurableExtensionDataRef<
-          RouteRef<AnyRouteRefParams>,
-          'core.routing.ref',
-          {
-            optional: true;
-          }
-        >
-      | ConfigurableExtensionDataRef<
-          string,
-          'core.title',
-          {
-            optional: true;
-          }
-        >
-      | ConfigurableExtensionDataRef<
-          IconElement,
-          'core.icon',
-          {
-            optional: true;
-          }
-        >,
-      {
-        singleton: false;
-        optional: false;
-        internal: false;
-      }
-    >;
-  };
-  kind: 'page';
-  name: undefined;
-  params: {
-    path: string;
-    title?: string;
-    icon?: IconElement;
-    loader?: () => Promise<JSX_2.Element>;
-    routeRef?: RouteRef;
-    noHeader?: boolean;
-  };
-}>;
-
-// (No @packageDocumentation comment for this package)
-```
