+++ workspaces/tech-radar/plugins/tech-radar/report-legacy.api.md
@@ -1,0 +1,62 @@
+## API Report File for "@backstage-community/plugin-tech-radar"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { ApiRef } from '@backstage/core-plugin-api';
+import { BackstagePlugin } from '@backstage/core-plugin-api';
+import { JSX as JSX_2 } from 'react/jsx-runtime';
+import { RouteRef } from '@backstage/core-plugin-api';
+import { TechRadarLoaderResponse } from '@backstage-community/plugin-tech-radar-common';
+
+// @public
+export function RadarPage(props: TechRadarPageProps): JSX_2.Element;
+
+// @public @deprecated (undocumented)
+export const Router: typeof RadarPage;
+
+// @public
+export interface TechRadarApi {
+  load: (id: string | undefined) => Promise<TechRadarLoaderResponse>;
+}
+
+// @public
+export const techRadarApiRef: ApiRef<TechRadarApi>;
+
+// @public
+export function TechRadarComponent(
+  props: TechRadarComponentProps,
+): JSX_2.Element;
+
+// @public
+export interface TechRadarComponentProps {
+  height?: number;
+  id?: string;
+  searchText?: string;
+  svgProps?: object;
+  width?: number;
+}
+
+// @public
+export const TechRadarPage: RadarPage;
+
+// @public
+export interface TechRadarPageProps extends TechRadarComponentProps {
+  pageTitle?: string;
+  subtitle?: string;
+  title?: string;
+}
+
+// @public
+const techRadarPlugin: BackstagePlugin<
+  {
+    root: RouteRef<undefined>;
+  },
+  {},
+  {}
+>;
+export { techRadarPlugin as plugin };
+export { techRadarPlugin };
+
+// (No @packageDocumentation comment for this package)
+```
