+++ workspaces/tech-radar/plugins/tech-radar/report.api.md
@@ -3,17 +3,130 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+import { AnyApiFactory } from '@backstage/frontend-plugin-api';
+import { AnyRouteRefParams } from '@backstage/frontend-plugin-api';
+import { ApiFactory } from '@backstage/frontend-plugin-api';
 import { ApiRef } from '@backstage/core-plugin-api';
-import { BackstagePlugin } from '@backstage/core-plugin-api';
-import { JSX as JSX_2 } from 'react/jsx-runtime';
-import { RouteRef } from '@backstage/core-plugin-api';
+import { ConfigurableExtensionDataRef } from '@backstage/frontend-plugin-api';
+import { ExtensionBlueprintParams } from '@backstage/frontend-plugin-api';
+import { ExtensionDataRef } from '@backstage/frontend-plugin-api';
+import { ExtensionInput } from '@backstage/frontend-plugin-api';
+import { IconElement } from '@backstage/frontend-plugin-api';
+import { JSX as JSX_2 } from 'react';
+import { OverridableExtensionDefinition } from '@backstage/frontend-plugin-api';
+import { OverridableFrontendPlugin } from '@backstage/frontend-plugin-api';
+import { RouteRef } from '@backstage/frontend-plugin-api';
 import { TechRadarLoaderResponse } from '@backstage-community/plugin-tech-radar-common';
 
 // @public
-export function RadarPage(props: TechRadarPageProps): JSX_2.Element;
-
-// @public @deprecated (undocumented)
-export const Router: typeof RadarPage;
+const _default: OverridableFrontendPlugin<
+  {
+    root: RouteRef<undefined>;
+  },
+  {},
+  {
+    'api:tech-radar': OverridableExtensionDefinition<{
+      kind: 'api';
+      name: undefined;
+      config: {};
+      configInput: {};
+      output: ExtensionDataRef<AnyApiFactory, 'core.api.factory', {}>;
+      inputs: {};
+      params: <
+        TApi,
+        TImpl extends TApi,
+        TDeps extends { [name in string]: unknown },
+      >(
+        params: ApiFactory<TApi, TImpl, TDeps>,
+      ) => ExtensionBlueprintParams<AnyApiFactory>;
+    }>;
+    'page:tech-radar': OverridableExtensionDefinition<{
+      config: {
+        subtitle: string;
+        pageTitle: string;
+        width: number | undefined;
+        height: number | undefined;
+        path: string | undefined;
+        title: string | undefined;
+      };
+      configInput: {
+        subtitle?: string | undefined;
+        pageTitle?: string | undefined;
+        width?: number | undefined;
+        height?: number | undefined;
+        path?: string | undefined | undefined;
+        title?: string | undefined | undefined;
+      };
+      output:
+        | ExtensionDataRef<JSX_2.Element, 'core.reactElement', {}>
+        | ExtensionDataRef<string, 'core.routing.path', {}>
+        | ExtensionDataRef<
+            RouteRef<AnyRouteRefParams>,
+            'core.routing.ref',
+            {
+              optional: true;
+            }
+          >
+        | ExtensionDataRef<
+            string,
+            'core.title',
+            {
+              optional: true;
+            }
+          >
+        | ExtensionDataRef<
+            IconElement,
+            'core.icon',
+            {
+              optional: true;
+            }
+          >;
+      inputs: {
+        pages: ExtensionInput<
+          | ConfigurableExtensionDataRef<JSX_2.Element, 'core.reactElement', {}>
+          | ConfigurableExtensionDataRef<string, 'core.routing.path', {}>
+          | ConfigurableExtensionDataRef<
+              RouteRef<AnyRouteRefParams>,
+              'core.routing.ref',
+              {
+                optional: true;
+              }
+            >
+          | ConfigurableExtensionDataRef<
+              string,
+              'core.title',
+              {
+                optional: true;
+              }
+            >
+          | ConfigurableExtensionDataRef<
+              IconElement,
+              'core.icon',
+              {
+                optional: true;
+              }
+            >,
+          {
+            singleton: false;
+            optional: false;
+            internal: false;
+          }
+        >;
+      };
+      kind: 'page';
+      name: undefined;
+      params: {
+        path: string;
+        title?: string;
+        icon?: IconElement;
+        loader?: () => Promise<JSX_2.Element>;
+        routeRef?: RouteRef;
+        noHeader?: boolean;
+      };
+    }>;
+  }
+>;
+export default _default;
 
 // @public
 export interface TechRadarApi {
@@ -21,40 +134,110 @@ export interface TechRadarApi {
 }
 
 // @public
+export const techRadarApi: OverridableExtensionDefinition<{
+  kind: 'api';
+  name: undefined;
+  config: {};
+  configInput: {};
+  output: ExtensionDataRef<AnyApiFactory, 'core.api.factory', {}>;
+  inputs: {};
+  params: <
+    TApi,
+    TImpl extends TApi,
+    TDeps extends { [name in string]: unknown },
+  >(
+    params: ApiFactory<TApi, TImpl, TDeps>,
+  ) => ExtensionBlueprintParams<AnyApiFactory>;
+}>;
+
+// @public
 export const techRadarApiRef: ApiRef<TechRadarApi>;
 
 // @public
-export function TechRadarComponent(
-  props: TechRadarComponentProps,
-): JSX_2.Element;
+export const techRadarPage: OverridableExtensionDefinition<{
+  config: {
+    subtitle: string;
+    pageTitle: string;
+    width: number | undefined;
+    height: number | undefined;
+    path: string | undefined;
+    title: string | undefined;
+  };
+  configInput: {
+    subtitle?: string | undefined;
+    pageTitle?: string | undefined;
+    width?: number | undefined;
+    height?: number | undefined;
+    path?: string | undefined | undefined;
+    title?: string | undefined | undefined;
+  };
+  output:
+    | ExtensionDataRef<JSX_2.Element, 'core.reactElement', {}>
+    | ExtensionDataRef<string, 'core.routing.path', {}>
+    | ExtensionDataRef<
+        RouteRef<AnyRouteRefParams>,
+        'core.routing.ref',
+        {
+          optional: true;
+        }
+      >
+    | ExtensionDataRef<
+        string,
+        'core.title',
+        {
+          optional: true;
+        }
+      >
+    | ExtensionDataRef<
+        IconElement,
+        'core.icon',
+        {
+          optional: true;
+        }
+      >;
+  inputs: {
+    pages: ExtensionInput<
+      | ConfigurableExtensionDataRef<JSX_2.Element, 'core.reactElement', {}>
+      | ConfigurableExtensionDataRef<string, 'core.routing.path', {}>
+      | ConfigurableExtensionDataRef<
+          RouteRef<AnyRouteRefParams>,
+          'core.routing.ref',
+          {
+            optional: true;
+          }
+        >
+      | ConfigurableExtensionDataRef<
+          string,
+          'core.title',
+          {
+            optional: true;
+          }
+        >
+      | ConfigurableExtensionDataRef<
+          IconElement,
+          'core.icon',
+          {
+            optional: true;
+          }
+        >,
+      {
+        singleton: false;
+        optional: false;
+        internal: false;
+      }
+    >;
+  };
+  kind: 'page';
+  name: undefined;
+  params: {
+    path: string;
+    title?: string;
+    icon?: IconElement;
+    loader?: () => Promise<JSX_2.Element>;
+    routeRef?: RouteRef;
+    noHeader?: boolean;
+  };
+}>;
 
-// @public
-export interface TechRadarComponentProps {
-  height?: number;
-  id?: string;
-  searchText?: string;
-  svgProps?: object;
-  width?: number;
-}
-
-// @public
-export const TechRadarPage: RadarPage;
-
-// @public
-export interface TechRadarPageProps extends TechRadarComponentProps {
-  pageTitle?: string;
-  subtitle?: string;
-  title?: string;
-}
-
-// @public
-const techRadarPlugin: BackstagePlugin<
-  {
-    root: RouteRef<undefined>;
-  },
-  {},
-  {}
->;
-export { techRadarPlugin as plugin };
-export { techRadarPlugin };
+// (No @packageDocumentation comment for this package)
 ```
