+++ workspaces/tech-radar/plugins/tech-radar/src/apis.ts
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2026 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {
+  ApiBlueprint,
+  discoveryApiRef,
+  fetchApiRef,
+  identityApiRef,
+} from '@backstage/frontend-plugin-api';
+import { techRadarApiRef } from './api';
+import { DefaultTechRadarApi } from './defaultApi';
+
+/**
+ * Tech Radar API extension for the new frontend system.
+ *
+ * @public
+ */
+export const techRadarApi = ApiBlueprint.make({
+  params: defineParams =>
+    defineParams({
+      api: techRadarApiRef,
+      deps: {
+        identityApi: identityApiRef,
+        discoveryApi: discoveryApiRef,
+        fetchApi: fetchApiRef,
+      },
+      factory: ({ identityApi, discoveryApi, fetchApi }) => {
+        return new DefaultTechRadarApi({
+          discoveryApi,
+          fetchApi,
+          identityApi,
+        });
+      },
+    }),
+});
