+++ workspaces/tech-radar/plugins/tech-radar/src/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2026 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,24 +20,26 @@
  * @packageDocumentation
  */
 
-import { RadarPage } from './components';
-
-export {
-  techRadarPlugin,
-  techRadarPlugin as plugin,
-  TechRadarPage,
-} from './plugin';
-
-export * from './components';
+import { convertLegacyRouteRefs } from '@backstage/core-compat-api';
+import { createFrontendPlugin } from '@backstage/frontend-plugin-api';
+import { techRadarApi } from './apis';
+import { techRadarPage } from './pages';
+import { rootRouteRef } from './plugin';
 
 /**
- * @deprecated Use plugin extensions instead
+ * Tech Radar plugin for the new frontend system.
  *
  * @public
  */
-export const Router = RadarPage;
+export default createFrontendPlugin({
+  pluginId: 'tech-radar',
+  extensions: [techRadarPage, techRadarApi],
+  routes: convertLegacyRouteRefs({
+    root: rootRouteRef,
+  }),
+});
 
-/**
- * The TypeScript API for configuring Tech Radar.
- */
-export * from './api';
+export { techRadarApi } from './apis';
+export { techRadarPage } from './pages';
+export { techRadarApiRef } from './api';
+export type { TechRadarApi } from './api';
