+++ workspaces/tech-radar/plugins/tech-radar/src/plugin.nfs.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 The Backstage Authors
+ * Copyright 2026 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,4 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import '@testing-library/jest-dom';
+
+import techRadarPlugin from '.';
+
+describe('Tech Radar plugin', () => {
+  describe('Plugin Structure', () => {
+    it('should have correct plugin metadata', () => {
+      expect(techRadarPlugin.pluginId).toBe('tech-radar');
+      expect(techRadarPlugin.routes.root).toBeDefined();
+    });
+
+    it('should export Tech Radar extensions', () => {
+      expect(techRadarPlugin.getExtension('api:tech-radar')).toBeDefined();
+      expect(techRadarPlugin.getExtension('page:tech-radar')).toBeDefined();
+    });
+  });
+});
